% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testLevel.R
\name{testLevel}
\alias{testLevel}
\title{testLevel}
\usage{
testLevel(x, xc)
}
\arguments{
\item{x}{a data vector of a single data set named and sorted by years}

\item{xc}{a data vector of a single data set named and sorted by years}
}
\value{
A named vector containing results of multiple tests
for the single combination of data and comparison data.
}
\description{
A collection of tests to compare the level of data with comparison data
}
\details{
Details about the employed tests can be found here:
\url{http://redmine.pik-potsdam.de/projects/x-intern/wiki/Traffic_light_validation}
}
\examples{

#create some test data
test <- 1:40; names(test) <- 1995:2034
x <- test[20:40]
xc <- test[1:30] + rnorm(30)

testLevel(x,xc)

}
\seealso{
\code{\link{TLstatistics}},\code{\link{trafficlight}}
}
\author{
Markus Bonsch, Jan Philipp Dietrich
}
