% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testTrend.R
\name{testTrend}
\alias{testTrend}
\title{testTrend}
\usage{
testTrend(x, xc, t = 40)
}
\arguments{
\item{x}{a data vector of a single data set named and sorted by years}

\item{xc}{a data vector of a single data set named and sorted by years}

\item{t}{window size in years for which the trend should be tested. It is 
tried to create a window -t/2 to t/2 around the start year for trend testing.
If this does not work the window can be moved in the range of -t to t around
the start year to find comparison data. Comparison data is shifted from the given
window to start year to start year + t for analysis}
}
\value{
A named vector containing results of multiple tests
for the single combination of data and comparison data.
}
\description{
A collection of tests to compare the trend of data with comparison data
}
\details{
Details about the employed tests can be found here:
\url{http://redmine.pik-potsdam.de/projects/x-intern/wiki/Traffic_light_validation}
}
\examples{

#create some test data
test <- 1:80; names(test) <- 1995:2074
x <- test
xc <- test + rnorm(30)
names(xc) <- 1925:2004

testOverlap(x,xc)

}
\seealso{
\code{\link{TLstatistics}},\code{\link{trafficlight}}
}
\author{
Markus Bonsch, Jan Philipp Dietrich
}
