% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLevaluate.R
\name{TLevaluate}
\alias{TLevaluate}
\title{TLevaluate}
\usage{
TLevaluate(TLstatistics)
}
\arguments{
\item{TLstatistics}{MAgPIE object containing statistical test results as
returned by \code{\link{TLstatistics}}.}
}
\value{
matrix containing agreement classes (0 = red, 1 = yellow, 2 =
green).
}
\description{
Function to evaluate statistical tests concerning model-data compatibility
with respect to predefined thresholds to determine agreement classes.
}
\details{
Details about the thresholds and the aggregation can be found here:
\url{https://redmine.pik-potsdam.de/projects/magpie-intern/wiki/Traffic_light_validation}
}
\examples{
library(magclass)
data("population_magpie")
x <- population_magpie
xc <- x + runif(length(x)[1],0,10^5)
tl <- TLstatistics(x[,,1],xc)

TLevaluate(tl)

}
\seealso{
\code{\link{TLstatistics}},\code{\link{trafficlight}},\code{\link{TLplot}}
}
\author{
Markus Bonsch
}
