% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLplot.R
\name{TLplot}
\alias{TLplot}
\title{TLplot}
\usage{
TLplot(mode, detailed = TRUE, linesize = 0.5)
}
\arguments{
\item{mode}{A matrix containing traffic light values
returned by \code{\link{TLevaluate}}.}

\item{detailed}{boolean deciding whether detailed information should be plotted or not.
In detailed mode each block represents a test group with the first traffic light showing
the overall performance in this group and the following traffic lights showing the results
of single tests.}

\item{linesize}{line size of the traffic light borders}
}
\value{
A ggplot object.
}
\description{
Function to plot the results of the automatic validation procedure as a
traffic light symbol
}
\examples{
library(magclass)
data("population_magpie")
x <- population_magpie
xc <- x + runif(length(x)[1],-10^5,10^5)
tl <- TLstatistics(x[,,1],xc)
tle <- TLevaluate(tl)

TLplot(tle)

}
\seealso{
\code{\link{TLevaluate}},\code{\link{TLstatistics}},\code{\link{trafficlight}}
}
\author{
Jan Philipp Dietrich, Markus Bonsch
}
