% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLstatistics.R
\name{TLstatistics}
\alias{TLstatistics}
\title{TLstatistics}
\usage{
TLstatistics(x, xc)
}
\arguments{
\item{x}{model output which should be tested in magclass format. Only a single element in data dimension is allowed}

\item{xc}{comparison data on which x should be tested (magclass format)}
}
\value{
A matrix containing results of multiple tests
for all available comparison datasets.
}
\description{
Calculates several statistical measures to compare model results with data.
}
\details{
Details about the employed tests can be found here:
\url{https://redmine.pik-potsdam.de/projects/magpie-intern/wiki/Traffic_light_validation}
}
\examples{

library(magclass)
data("population_magpie")
x <- population_magpie
xc <- x + runif(length(x)[1],0,10^5)
TLstatistics(x[,,1],xc)

}
\seealso{
\code{\link{TLevaluate}},\code{\link{trafficlight}},\code{\link{TLplot}}
}
\author{
Markus Bonsch, Jan Philipp Dietrich
}
