% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafficlight-package.R
\docType{package}
\name{trafficlight-package}
\alias{trafficlight-package}
\title{Traffic light - Tools for data validation}
\description{
The package contains tools for data validation and aggregation of validation
results.
}
\details{
\tabular{ll}{ Package: \tab trafficlight\cr Type: \tab Package\cr Version:
\tab 1.000\cr Date: \tab 2016-10-07 \cr License: \tab LGPL-3\cr }
}
\author{
Markus Bonsch, Jan Philipp Dietrich

Maintainer: Jan Philipp Dietrich <dietrich@pik-potsdam.de>
}
