% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafficlight.R
\name{trafficlight}
\alias{trafficlight}
\title{trafficlight}
\usage{
trafficlight(x, xc, plot = TRUE, ...)
}
\arguments{
\item{x}{model output which should be tested in magclass format}

\item{xc}{comparison data on which x should be tested (magclass format)}

\item{plot}{TRUE or FALSE. If TRUE, a plot with a visualization of te
treffic light is returned. If FALSE, a MAgPIE object with the test result is
returned.}

\item{...}{Additional arguments provided to \code{\link{TLplot}}}
}
\value{
If \code{plot==TRUE}, a ggplot object. If \code{plot==FALSE}, a
matrix with traffic light result (0 = red, 1 = yellow, 2 = green)
}
\description{
Function to perform an automatic validation routine
}
\details{
The procedure is explained in detail here:
\url{https://redmine.pik-potsdam.de/projects/magpie-intern/wiki/Traffic_light_validation}
}
\examples{
library(magclass)
data("population_magpie")
x <- population_magpie
xc <- x + runif(length(x)[1],0,10^5)

trafficlight(x[,,1], xc)
trafficlight(x[,,1], xc, detailed=FALSE)
}
\seealso{
\code{\link{TLevaluate}},\code{\link{TLstatistics}},\code{\link{TLplot}}
}
\author{
Markus Bonsch, Jan Philipp Dietrich, Benjamin Leon Bodirsky
}
